% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_by_blocks.R
\name{apply_by_block}
\alias{apply_by_block}
\title{Modify matrix by blocks}
\usage{
apply_by_block(Y, layout, L, FUN, ...)
}
\arguments{
\item{Y}{Matrix}

\item{layout}{Blocks are distributed: 1 horizontally, 2 vertically.}

\item{L}{Number of lines of the block.}

\item{FUN}{Funtion to apply to the block.}

\item{...}{Arguments to be passed to FUN.}
}
\value{
A matrix.
}
\description{
Applies a function to a matrix by horizontal or vertical blocks.
}
