% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuadramelo-package.R
\docType{package}
\name{cuadramelo-package}
\alias{cuadramelo}
\alias{cuadramelo-package}
\title{cuadramelo: Matrix Balancing and Rounding}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Balancing and rounding matrices subject to restrictions. Adjustment of matrices so that columns and rows add up to given vectors, rounding of a matrix while keeping the column and/or row totals, performing these by blocks...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mserrano-ine.github.io/cuadramelo/}
}

}
\author{
\strong{Maintainer}: Miguel Serrano \email{miguel.serrano.martin@ine.es}

}
\keyword{internal}
