% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_by_blocks.R
\name{round_by_blocks}
\alias{round_by_blocks}
\title{Round matrix by blocks}
\usage{
round_by_blocks(Y, MARGIN, L, digits = 0)
}
\arguments{
\item{Y}{Matrix}

\item{MARGIN}{1 horizontal, 2 vertical.}

\item{L}{Number of lines of the block.}

\item{digits}{Number of decimal places to be rounded to.}
}
\description{
Applies \code{round_matrix()} to the blocks using \code{apply_by_block()}.
}
