% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_matrix.R
\name{round_matrix}
\alias{round_matrix}
\title{Round a matrix}
\usage{
round_matrix(mat, digits = 0)
}
\arguments{
\item{mat}{A matrix}

\item{digits}{Number of decimal places.}
}
\value{
A rounded matrix.
}
\description{
Rounds the entries of a matrix to \code{digits} decimal places while
preserving the rounded colSums and rowSums.
}
