% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_matrix.R
\name{round_vector}
\alias{round_vector}
\title{Round univariate}
\usage{
round_vector(x, digits = 0)
}
\arguments{
\item{x}{A vector.}

\item{digits}{Number of decimal places to be rounded to.}
}
\description{
Rounds a vector preserving the rounded sum.
}
