% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_matrix.R
\name{stack.matrix}
\alias{stack.matrix}
\title{Stack matrix columns}
\usage{
\method{stack}{matrix}(X)
}
\arguments{
\item{X}{A numerical matrix.}
}
\value{
A vector whose i*j-th element is the i-th
element of the j-th column.
}
\description{
Stack the columns of a matrix into a vector.
}
