% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_matrix.R
\name{balance_matrix}
\alias{balance_matrix}
\title{Balance matrix}
\usage{
balance_matrix(Y, v = NULL, h = NULL)
}
\arguments{
\item{Y}{Matrix to be balanced.}

\item{v}{(optional) Desired sum of columns.}

\item{h}{(optional) Desired sum of rows.}
}
\value{
A list containing \code{X}, the balanced matrix, and a
measure of the change the balancing did to the original matrix.
}
\description{
Balances a matrix so that the columns and/or rows add up
to a certain vector.
}
\details{
Balancing is done according to the criteria of minimum sum
of squares.

If neither \code{v} nor \code{h} is given, the same matrix will be
returned. If only one of them is given, only that axis will be
balanced.
}
\examples{
set.seed(2)
Y <- rnorm(3*5) |> matrix(3,5) |> round(3)
v <- c( 0.876, -1.078, 3.452, 0.261, 1.349)
h <- c(-1.851, 0.243, 6.468)
Y
X1 <- balance_matrix(Y, v, h)$X
h
rowSums(X1)
v
colSums(X1)
print(paste("The change has norm2 =", balance_matrix(Y, v, h)$change))
X2 <- balance_matrix(Y, v = v)
v
colSums(X2)
X3 <- balance_matrix(Y, h = h)
h
rowSums(X3)
}
