% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_matrix.R
\name{bg}
\alias{bg}
\title{Backtracking algorithm for rounding}
\usage{
bg(idx, row_diffs, col_diffs, elementos_marcados = list())
}
\description{
Solves the problem of rounding a matriz while preserving rowSums and colSums
via a backtracking algorithm.
}
