% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuadramelo-package.R
\docType{package}
\name{cuadramelo-package}
\alias{cuadramelo}
\alias{cuadramelo-package}
\title{cuadramelo: Matrix balancing and rounding}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package has functions for balancing and rounding matrices subject to restrictions. One can adjust the matrix so that columns and rows add up to given vectors, round a matrix while keeping the column and row totals, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mserrano-ine.github.io/cuadramelo/}
}

}
\author{
\strong{Maintainer}: Miguel Serrano \email{miguel.serrano.martin@ine.es}

}
\keyword{internal}
